feat: 퀴즈 제출 api 개발(LEAR-189)

+++ src/main/java/com/kkamjidot/api/mono/dto/response/ThisWeekResponse.java
@@ -4,6 +4,7 @@ import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.Builder;
 import lombok.Getter;
 
+import javax.persistence.criteria.CriteriaBuilder;
 import java.io.Serializable;
 import java.time.LocalDateTime;
 
@@ -11,7 +12,7 @@ import java.time.LocalDateTime;
 @Builder
 @Schema(name = "현재 주차 반환 응답")
 public class ThisWeekResponse implements Serializable {
-    private final Long week;
+    private final Integer week;
     private final Long challengeId;
     private final LocalDateTime challStartDate;
     private final LocalDateTime now;
